fix v1.1.1: added changes mentioned in the README, almost nuked the commits, avoided that fate barely.
