+++ .gitignore
@@ -30,3 +30,6 @@
 *.exe
 *.out
 *.app
+
+# Visual Studio
+*.vs
