+++ ardExample_alg01/alg01.cpp
@@ -3,5 +3,3 @@
 int Rounder::round(double num) {
     return static_cast<int>(std::round(num));
 }
-
-#include "export_to_pyd.h"
