+++ src/driver.cpp
@@ -1,7 +1,8 @@
 #include "alg01.h"
+#include<iostream>
 
 int main() {
     Rounder r;
     double num = 5.223;
-    r.round(num);
+    std::cout << r.round(num) << std::endl;
 }
